+++ pmd-javascript/src/test/java/net/sourceforge/pmd/lang/ecmascript/ast/JsTreeDumpTest.java
@@ -30,4 +30,9 @@ public class JsTreeDumpTest extends BaseTreeDumpTest {
     public void jquerySelector() {
         doTest("jquery-selector");
     }
+
+    @Test
+    public void decorators() {
+        doTest("decorators");
+    }
 }